pyproject.toml
[project]
name = "hotelrecommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "groq>=0.28.0",
    "langchain>=0.3.25",
    "langchain-groq>=0.3.2",
    "langchain-postgres>=0.0.14",
    "openai>=1.88.0",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "sqlalchemy>=2.0.41",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.3",
]

[tool.ruff]
select = ["E", "F", "I"]
ignore = ["F401", "F841", "F403", "F405"]
